requirements.txt
streamlit
torch
torchvision
torchcam
Pillow
requests
